% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_to_pdf}
\alias{convert_to_pdf}
\title{Convert to PDF}
\usage{
convert_to_pdf(tr_id, cv_ui)
}
\arguments{
\item{cv_ui}{}
}
\value{

}
\description{
Creates a shiny ui page and creates a snapshot in pdf,
named ...$tr_id.
cv_ui <- fluidPage(...)
}
