% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_tr_id}
\alias{generate_tr_id}
\title{Generate Transaction Id}
\usage{
generate_tr_id()
}
\value{

}
\description{
Id is an alphanumeric character string, length 8.
Used to store customer data and name files
}
