% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mail_resume}
\alias{mail_resume}
\title{Resume mailer}
\usage{
mail_resume(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Mail resume to client, ... is user info list
...$first_name is required
}
