% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{multiple_details_handler}
\alias{multiple_details_handler}
\title{Multiple Details Handler}
\usage{
multiple_details_handler(df, index_col)
}
\arguments{
\item{index_col}{}
}
\value{

}
\description{
Use to process worksheets that may contain extra rows for
the same item, index_col. Splits worksheet into a list of dataframes,
containing only one item from the index_col per dataframe
}
