% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_achievements_ws}
\alias{process_achievements_ws}
\title{Process achievements_ws}
\usage{
process_achievements_ws(achievements_ws)
}
\arguments{
\item{achievements_ws}{}
}
\value{

}
\description{
Call from processor to process achievements_ws
}
