% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_awards_ws}
\alias{process_awards_ws}
\title{Process awards_ws}
\usage{
process_awards_ws(awards_ws)
}
\arguments{
\item{awards_ws}{}
}
\value{

}
\description{
Call from processor to process awards_ws
}
