% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_boards_ws}
\alias{process_boards_ws}
\title{Process boards_ws}
\usage{
process_boards_ws(boards_ws)
}
\arguments{
\item{boards_ws}{}
}
\value{

}
\description{
Call from processor to process boards_ws
}
