% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_education_ws}
\alias{process_education_ws}
\title{Process education_ws}
\usage{
process_education_ws(education_ws)
}
\arguments{
\item{education_ws}{}
}
\value{

}
\description{
Call from processor to process education_ws
}
