% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_experience_ws}
\alias{process_experience_ws}
\title{Process experience_ws}
\usage{
process_experience_ws(experience_ws)
}
\arguments{
\item{experience_ws}{}
}
\value{

}
\description{
Call from processor to process experience_ws
}
