% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_grants_ws}
\alias{process_grants_ws}
\title{Process grants_ws}
\usage{
process_grants_ws(grants_ws)
}
\arguments{
\item{grants_ws}{}
}
\value{

}
\description{
Call from processor to process grants_ws
}
