% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_history_ws}
\alias{process_history_ws}
\title{Process history_ws}
\usage{
process_history_ws(history_ws)
}
\arguments{
\item{history_ws}{}
}
\value{

}
\description{
Call from processor to process history_ws
}
