% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_memberships_ws}
\alias{process_memberships_ws}
\title{Process memberships_ws}
\usage{
process_memberships_ws(memberships_ws)
}
\arguments{
\item{memberships_ws}{}
}
\value{

}
\description{
Call from processor to process memberships_ws
}
