% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_personal_ws}
\alias{process_personal_ws}
\title{Process personal_ws}
\usage{
process_personal_ws(personal_ws)
}
\arguments{
\item{personal_ws}{}
}
\value{

}
\description{
Call from processor to process personal_ws
}
