% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_references_ws}
\alias{process_references_ws}
\title{Process references_ws}
\usage{
process_references_ws(references_ws)
}
\arguments{
\item{references_ws}{}
}
\value{

}
\description{
Call from processor to process references_ws
}
