% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_reports_ws}
\alias{process_reports_ws}
\title{Process reports_ws}
\usage{
process_reports_ws(reports_ws)
}
\arguments{
\item{reports_ws}{}
}
\value{

}
\description{
Call from processor to process reports_ws
}
