% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_research_ws}
\alias{process_research_ws}
\title{Process research_ws}
\usage{
process_research_ws(research_ws)
}
\arguments{
\item{research_ws}{}
}
\value{

}
\description{
Call from processor to process research_ws
}
