% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_services_ws}
\alias{process_services_ws}
\title{Process services_ws}
\usage{
process_services_ws(services_ws)
}
\arguments{
\item{services_ws}{}
}
\value{

}
\description{
Call from processor to process services_ws
}
