% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_skills_2_ws}
\alias{process_skills_2_ws}
\title{Process skills_2_ws}
\usage{
process_skills_2_ws(skills_2_ws)
}
\arguments{
\item{skills_2_ws}{}
}
\value{

}
\description{
Call from processor to process skills_2_ws
}
