% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_skills_ws}
\alias{process_skills_ws}
\title{Process skills_ws}
\usage{
process_skills_ws(skills_ws)
}
\arguments{
\item{skills_ws}{}
}
\value{

}
\description{
Call from processor to process skills_ws
}
