% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_software_ws}
\alias{process_software_ws}
\title{Process software_ws}
\usage{
process_software_ws(software_ws)
}
\arguments{
\item{software_ws}{}
}
\value{

}
\description{
Call from processor to process software_ws
}
