% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_solutions_ws}
\alias{process_solutions_ws}
\title{Process solutions_ws}
\usage{
process_solutions_ws(solutions_ws)
}
\arguments{
\item{solutions_ws}{}
}
\value{

}
\description{
Call from processor to process solutions_ws
}
