% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_talks_ws}
\alias{process_talks_ws}
\title{Process talks_ws}
\usage{
process_talks_ws(talks_ws)
}
\arguments{
\item{talks_ws}{}
}
\value{

}
\description{
Call from processor to process talks_ws
}
