% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_teaching_ws}
\alias{process_teaching_ws}
\title{Process teaching_ws}
\usage{
process_teaching_ws(teaching_ws)
}
\arguments{
\item{teaching_ws}{}
}
\value{

}
\description{
Call from processor to process teaching_ws
}
