% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_thesis_ws}
\alias{process_thesis_ws}
\title{Process thesis_ws}
\usage{
process_thesis_ws(thesis_ws)
}
\arguments{
\item{thesis_ws}{}
}
\value{

}
\description{
Call from processor to process thesis_ws
}
