% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{process_user_data}
\alias{process_user_data}
\title{Process User Data}
\usage{
process_user_data(user_data)
}
\arguments{
\item{user_data}{}
}
\value{

}
\description{
Get user data from list of worksheets ready for templates
}
