% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ws.R
\name{processor}
\alias{processor}
\title{Processor}
\usage{
processor(worksheets, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
switch case statement to run process_ on corresponding worksheet.
Map over names of worksheets
}
