% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_rownames_column}
\alias{remove_rownames_column}
\title{Delete rownames column}
\usage{
remove_rownames_column(worksheet)
}
\arguments{
\item{worksheet}{}
}
\value{

}
\description{
Use in a case where worksheet is read with an extra rownames column
}
