% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_csvs}
\alias{split_csvs}
\title{Split by commas}
\usage{
split_csvs(table_cell)
}
\arguments{
\item{table_cell}{}
}
\value{

}
\description{
Use in worksheet sections where multiple values should be split by commas
}
