% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ss_builder}
\alias{ss_builder}
\title{Spreadsheet builder}
\usage{
ss_builder(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Takes in user list containing info on current user, named ...$tr_id
Subsets ...$user_template over templates_data and returns worksheets selected in
a list, creates an xlsx file named ...$tr_id in temp folder, mails it to
...$email and then deletes the xlsx file
...$first_name and ...$method are required
...$method = "mail", include "googlesheets" maybe or remove
}
